factory combinator

+++ Core/Point.cs
@@ -1,3 +1,3 @@
 ﻿namespace Core;
 
-public record Point(int X, int Y);
+public record Point(double X, double Y);
